fix: file upload service

+++ packages/common/src/services/file-upload.service.ts
@@ -1,10 +1,18 @@
-import { FastifyInstance } from 'fastify';
 import * as fs from 'fs';
 import * as path from 'path';
-import * as fastify from 'fastify';
-import { Injectable, InternalServerErrorException, Logger } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  InternalServerErrorException,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { Client } from 'minio';
-import { MultipartFile, STORAGE_MODE } from '../interfaces/file-upload.interface';
+import { encode } from 'urlencode';
+import {
+  MultipartFile,
+  STORAGE_MODE,
+} from '../interfaces/file-upload.interface';
 import {
   FileDownloadRequestDTO,
   FileUploadRequestDTO,
@@ -16,31 +24,58 @@ import { ConfigService } from '@nestjs/config';
 @Injectable()
 export class FileUploadService {
   private readonly storage: any;
-  private readonly useService: boolean;
-  private readonly fastifyInstance: FastifyInstance;
+  private readonly storageMode: string;
   private readonly useSSL: boolean;
   private readonly storageEndpoint: string;
   private readonly storagePort: number;
-  private readonly bucketName: string;
   private logger: Logger;
 
   constructor(private readonly configService: ConfigService) {
-    this.logger = new Logger('FileUploadService');
-    this.useService = this.configService.get<string>('STORAGE_MODE')?.toLowerCase() === STORAGE_MODE.MINIO;
+    this.logger = new Logger(FileUploadService.name);
+    this.storageMode = this.configService
+      .get<string>('STORAGE_MODE')
+      ?.toLowerCase();
     this.useSSL = this.configService.get<string>('STORAGE_USE_SSL') === 'true';
     this.storageEndpoint = this.configService.get<string>('STORAGE_ENDPOINT');
     this.storagePort = parseInt(this.configService.get('STORAGE_PORT'), 10);
-    this.bucketName = this.configService.get<string>('MINIO_BUCKETNAME');
-    if (this.useService) {
-      this.storage = new Client({
-        endPoint: this.storageEndpoint,
-        port: this.storagePort,
-        useSSL: this.useSSL,
-        accessKey: this.configService.get('STORAGE_ACCESS_KEY'),
-        secretKey: this.configService.get('STORAGE_SECRET_KEY'),
-      });
+
+    switch (this.storageMode) {
+      case STORAGE_MODE.MINIO:
+        this.logger.verbose('using minio');
+        this.storage = new Client({
+          endPoint: this.storageEndpoint,
+          port: this.storagePort,
+          useSSL: this.useSSL,
+          accessKey: this.configService.get('STORAGE_ACCESS_KEY'),
+          secretKey: this.configService.get('STORAGE_SECRET_KEY'),
+        });
+        break;
+      case STORAGE_MODE.LOCAL:
+        this.logger.verbose('using local storage.');
+        if (!fs.existsSync(this.storageEndpoint)) {
+          throw new InternalServerErrorException(
+            'Storage location does not exist. Make sure this location is present and accessible by the application.',
+          );
+        }
+      default:
+        break;
+    }
+  }
+
+  private buildURL(destination: string, filename: string): string {
+    const localhostRegex = /localhost/;
+    const ipv4Regex =
+      /\b((25[0-5]|2[0-4][0-9]|1[0-9]{2}|[1-9]?[0-9])\.){3}(25[0-5]|2[0-4][0-9]|1[0-9]{2}|[1-9]?[0-9])\b/;
+    const ipv6Regex =
+      /\b((25[0-5]|2[0-4][0-9]|1[0-9]{2}|[1-9]?[0-9])\.){3}(25[0-5]|2[0-4][0-9]|1[0-9]{2}|[1-9]?[0-9])\b|\b(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:((:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]|)[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]|)[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]|)[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]|)[0-9]))\b/;
+    if (
+      localhostRegex.test(this.storageEndpoint) ||
+      ipv4Regex.test(this.storageEndpoint) ||
+      ipv6Regex.test(this.storageEndpoint)
+    ) {
+      return `${this.useSSL ? 'https' : 'http'}://${this.storageEndpoint}:${this.storagePort}/${destination}/${encode(filename)}`;
     } else {
-      this.fastifyInstance = fastify();
+      return `${this.useSSL ? 'https' : 'http'}://${this.storageEndpoint}/${destination}/${encode(filename)}`;
     }
   }
 
@@ -48,10 +83,22 @@ export class FileUploadService {
     const metaData = {
       'Content-Type': uploadToMinioRequestDto.file.mimetype,
     };
+    const { destination, filename } = uploadToMinioRequestDto;
+    let exists = false;
+    try {
+      exists = await this.storage.bucketExists(destination);
+    } catch (err) {
+      this.logger.error(`Error checking bucket presence: ${err.message}`);
+    }
+
+    if (!exists) {
+      throw new BadRequestException('Bucket does not exist please create one.');
+    }
+
     return new Promise((resolve, reject) => {
       this.storage.putObject(
-         this.configService.get<string>('MINIO_BUCKETNAME'), 
-        uploadToMinioRequestDto.filename,
+        destination,
+        filename,
         uploadToMinioRequestDto.file.buffer,
         metaData,
         (err) => {
@@ -59,9 +106,7 @@ export class FileUploadService {
             this.logger.error(`Error uploading to Minio: ${err.message}`);
             reject(err);
           }
-          resolve(
-            `${this.useSSL ? 'https' : 'http'}://${this.storageEndpoint}:${this.storagePort}/${this.bucketName}/${uploadToMinioRequestDto.filename}`,
-          );
+          resolve(this.buildURL(destination, filename));
         },
       );
     });
@@ -79,21 +124,24 @@ export class FileUploadService {
         throw new InternalServerErrorException('File upload failed: directory creation error');
       }
     } else {
-      this.logger.log(`Directory already exists at ${uploadsDir}`);
+      throw new BadRequestException(
+        'Given destination path does not exist. Please create one.',
+      );
     }
-    fs.writeFileSync(localFilePath, saveToLocalRequestDto.file.buffer);
-    return saveToLocalRequestDto.destination;
   }
 
   async upload(fileUploadRequestDto : FileUploadRequestDTO): Promise<string> {
     try {
-      switch (this.configService.get<string>('STORAGE_MODE')?.toLowerCase()) {
-        case STORAGE_MODE.MINIO:  
+      switch (this.storageMode) {
+        case STORAGE_MODE.MINIO:
           this.logger.log('using minio');
           return await this.uploadToMinio(fileUploadRequestDto);
-        default:
+        case STORAGE_MODE.LOCAL:
           this.logger.log('writing to storage');
           return await this.saveLocalFile(fileUploadRequestDto);
+        default:
+          this.logger.error('Invalid storage mode');
+          throw new InternalServerErrorException('Invalid storage mode');
       }
     } catch (error) {
       this.logger.error(`Error uploading file: ${error.message}`);
